requirements.txt
# SPDX-FileCopyrightText: 2025 Bachibouzouk <pierre-francois.duc@rl-institute.de>
# SPDX-FileCopyrightText: 2025 Bryan Lancien <bryanlancien@yahoo.fr>
# SPDX-FileCopyrightText: 2025 Christian Winger <c@wingechr.de>
# SPDX-FileCopyrightText: 2025 Johann Wagner <johann@wagnerdevelopment.de>
# SPDX-FileCopyrightText: 2025 Jonas H <jonas.huber@rl-institut.de>
# SPDX-FileCopyrightText: 2025 Jonas Huber <38939526+jh-RLI@users.noreply.github.com>
# SPDX-FileCopyrightText: 2025 Jonas Huber <jonas.huber@rl-institut.de>
# SPDX-FileCopyrightText: 2025 Kirann Bhavaraju <kirann9999@hotmail.com>
# SPDX-FileCopyrightText: 2025 MGlauer <martinglauer89@gmail.com>
# SPDX-FileCopyrightText: 2025 MGlauer <martinglauer89@googlemail.com>
# SPDX-FileCopyrightText: 2025 Martin Glauer <martinglauer89@gmail.com>
# SPDX-FileCopyrightText: 2025 Martin Glauer <martinglauer89@googlemail.com>
# SPDX-FileCopyrightText: 2025 Tu Phan Ngoc <RL-INSTITUT\tuphan.ngoc@rli-nb-65.rl-institut.local>
# SPDX-FileCopyrightText: 2025 Venkatesh Murugadas <venkatesh.murugadas@st.ovgu.de>
# SPDX-FileCopyrightText: 2025 Zoë Schaaff <z.schaaff@oeko.de>
# SPDX-FileCopyrightText: 2025 c.winger <c.winger@oeko.de>
# SPDX-FileCopyrightText: 2025 christian-rli <christian.hofmann@rl-institut.de>
# SPDX-FileCopyrightText: 2025 daryna <Daryna.Martymianova@rl-institut.de>
# SPDX-FileCopyrightText: 2025 dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
# SPDX-FileCopyrightText: 2025 jh-RLI <38939526+jh-RLI@users.noreply.github.com>
# SPDX-FileCopyrightText: 2025 jh-RLI <jonas.huber@rl-institut.de>
# SPDX-FileCopyrightText: 2025 pietro <peter.zamb@gmail.com>
# SPDX-FileCopyrightText: 2025 solar-c <@>
# SPDX-FileCopyrightText: 2025 wingechr <wingechr@users.noreply.github.com>
#
# SPDX-License-Identifier: CC0-1.0

Django==5.1.4
matplotlib==3.8.3
urllib3==1.26.19
# scipy==1.12.0 # big package, only used when attemping to get the github activity graph -> not implemented ATM
django-widget-tweaks==1.5.0
SQLAlchemy==1.3.24 #  <1.4 is important (test_sync_tags_keywords_969, test_issue_832_tag_filter_with_missing_tag)
Pillow==10.3.0
bibtexparser==1.4.1
django-colorfield==0.7.3
requests==2.32.0
psycopg2-binary==2.9.9
webcolors==1.13
djangorestframework==3.15.2
django-axes==5.41.1
Shapely==1.8.5.post1 # In python v3.12 use at least Shapely==2.0.0
markdown2==2.4.13
alembic==1.13.1
django-simple-captcha==0.5.20
django-bootstrap5==21.3
django-fontawesome-5==1.0.18
omi==1.1.0
zipstream-new==1.1.8
rdflib==7.0.0
pyparsing==2.4.7 # <3 otherwise: sparql queries as text cannot be parsed
django-better-admin-arrayfield==1.4.2
SPARQLWrapper==1.8.5
GeoAlchemy2==0.11.1
jsonschema==4.22.0 # Mainly required by omi / oemetadata
oemetadata==2.0.4
django-compressor==4.5.1
django-cors-headers==4.3.1
owlready2==0.45
